Refactor FD connection logic into a new utility module; streamline connection handling in EditorCanvas
